RLInv/pyproject.toml
[project]
name = "rlinv"
version = "0.1.0"
description = "Loop Invariant Generation using Reinforcement Learning"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    {name = "Ido Pinto", email = "ido.pinto@mail.huji.ac.il"},
]
keywords = ["loop-invariants", "verification", "reinforcement-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
]

# Core dependencies (always required)
dependencies = [
    "datasets>=4.4.0",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.7",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "peft>=0.17.1",
    "pycparser>=2.23",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.3",
    "seaborn>=0.13.2",
    "together>=1.5.26",
    "tqdm>=4.67.1",
    "trackio>=0.8.1",
    "trl>=0.24.0",
    "wandb>=0.22.3",
    "weave>=0.52.16",
    "z3-solver>=4.15.4.0",
]

# Optional dependencies for different features
[project.optional-dependencies]
ai = [
    "openai>=2.5.0",
    "together>=1.5.26",
]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=1.0.0",
]
all = [
    "rlinv[ai,dev]",
]

# Console scripts
[project.scripts]
rlinv-eval = "src.eval.evaluate:main"

# Build system
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Package configuration
[tool.hatch.build.targets.wheel]
packages = ["src"]

# Tool configurations
[tool.black]
line-length = 88
target-version = ['py313']

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
